Inline single feature check
